requirements.txt
asgiref==3.5.1
dj-database-url==0.5.0
Django==4.0.4
django-environ==0.8.1
django-heroku==0.3.1
djangorestframework==3.13.1
Dynamic-Learning-Technique==0.2.2
GPUtil==1.4.0
joblib==1.1.0
numpy==1.22.3
pandas==1.4.2
psutil==5.9.0
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tzdata==2022.1
whitenoise==6.0.0
xgboost~=1.6.1
gunicorn